Revert "chore: migrate stats to otel (#2795)"
This reverts commit e839900d259c845a669ad743361099fa604af666.
