+++ .github/workflows/verify.yml
@@ -61,4 +61,4 @@ jobs:
       - name: golangci-lint
         uses: golangci/golangci-lint-action@v3
         with:
-          version: v1.50.1
+          version: v1.49.0
