+++ app/apphandlers/setup.go
@@ -26,7 +26,7 @@ import (
 type AppHandler interface {
 	// Setup to be called only once before starting the app.
 	Setup(*app.Options) error
-	// StartRudderCore starts the app
+	// Start starts the app
 	StartRudderCore(context.Context, *app.Options) error
 }
