+++ jobsdb/admin.go
@@ -21,7 +21,7 @@ func (jd *HandleT) Status() interface{} {
 	statusObj["empty-results-cache"] = emptyResults
 
 	pendingEventMetrics := metric.Instance.
-		GetRegistry(metric.PublishedMetrics).
+		GetRegistry(metric.PUBLISHED_METRICS).
 		GetMetricsByName(fmt.Sprintf(metric.JOBSDB_PENDING_EVENTS_COUNT, jd.tablePrefix))
 
 	if len(pendingEventMetrics) == 0 {
