+++ services/stats/memstats/stats_test.go
@@ -5,10 +5,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/require"
-
 	"github.com/rudderlabs/rudder-server/services/stats"
 	"github.com/rudderlabs/rudder-server/services/stats/memstats"
+	"github.com/stretchr/testify/require"
 )
 
 func TestStats(t *testing.T) {
@@ -127,7 +126,7 @@ func TestStats(t *testing.T) {
 	})
 
 	t.Run("no op", func(t *testing.T) {
-		require.NoError(t, store.Start(context.Background()))
+		store.Start(context.Background())
 		store.Stop()
 	})
