+++ services/stats/periodic.go
@@ -1,7 +1,6 @@
 package stats
 
 import (
-	"context"
 	"runtime"
 	"time"
 
@@ -13,16 +12,6 @@ import (
 // parts separated by the '.' character in the form used by statsd (e.x.
 // "mem.heap.alloc")
 type gaugeFunc func(key string, val uint64)
-
-// periodicStatsConfig is the configuration for the periodic stats collection
-type periodicStatsConfig struct {
-	enabled                 bool
-	statsCollectionInterval int64
-	enableCPUStats          bool
-	enableMemStats          bool
-	enableGCStats           bool
-	metricManager           metric.Manager
-}
 
 // runtimeStatsCollector implements the periodic grabbing of informational data from the
 // runtime package and outputting the values to a GaugeFunc.
@@ -66,8 +55,7 @@ func newRuntimeStatsCollector(gaugeFunc gaugeFunc) runtimeStatsCollector {
 // Run gathers statistics from package runtime and outputs them to the configured GaugeFunc every
 // PauseDur. This function will not return until Done has been closed (or never if Done is nil),
 // therefore it should be called in its own goroutine.
-func (c runtimeStatsCollector) run(ctx context.Context) {
-	defer close(c.done)
+func (c runtimeStatsCollector) run() {
 	defer c.zeroStats()
 	c.outputStats()
 
@@ -77,7 +65,7 @@ func (c runtimeStatsCollector) run(ctx context.Context) {
 	defer tick.Stop()
 	for {
 		select {
-		case <-ctx.Done():
+		case <-c.done:
 			return
 		case <-tick.C:
 			c.outputStats()
@@ -192,8 +180,7 @@ func newMetricStatsCollector(stats Stats, metricManager metric.Manager) metricSt
 }
 
 // run gathers statistics from package metric and outputs them as
-func (c metricStatsCollector) run(ctx context.Context) {
-	defer close(c.done)
+func (c metricStatsCollector) run() {
 	c.outputStats()
 
 	// Gauges are a 'snapshot' rather than a histogram. Pausing for some interval
@@ -202,7 +189,7 @@ func (c metricStatsCollector) run(ctx context.Context) {
 	defer tick.Stop()
 	for {
 		select {
-		case <-ctx.Done():
+		case <-c.done:
 			return
 		case <-tick.C:
 			c.outputStats()
@@ -214,17 +201,17 @@ func (c metricStatsCollector) outputStats() {
 	if c.metricManager == nil {
 		return
 	}
-	c.metricManager.GetRegistry(metric.PublishedMetrics).Range(func(key, value interface{}) bool {
+	c.metricManager.GetRegistry(metric.PUBLISHED_METRICS).Range(func(key, value interface{}) bool {
 		m := key.(metric.Measurement)
 		switch value := value.(type) {
 		case metric.Gauge:
-			c.stats.NewTaggedStat(m.GetName(), GaugeType, m.GetTags()).
+			c.stats.NewTaggedStat(m.GetName(), GaugeType, Tags(m.GetTags())).
 				Gauge(value.Value())
 		case metric.Counter:
-			c.stats.NewTaggedStat(m.GetName(), CountType, m.GetTags()).
+			c.stats.NewTaggedStat(m.GetName(), CountType, Tags(m.GetTags())).
 				Count(int(value.Value()))
 		case metric.MovingAverage:
-			c.stats.NewTaggedStat(m.GetName(), GaugeType, m.GetTags()).
+			c.stats.NewTaggedStat(m.GetName(), GaugeType, Tags(m.GetTags())).
 				Gauge(value.Value())
 		}
 		return true
