--- services/stats/testdata/otel-collector-config.yaml
@@ -1,24 +1,0 @@
-receivers:
-  otlp:
-    protocols:
-      grpc:
-
-exporters:
-  prometheus:
-    endpoint: "0.0.0.0:8889"
-    const_labels:
-      label1: value1
-
-processors:
-  batch:
-
-extensions:
-  health_check:
-
-service:
-  extensions: [health_check]
-  pipelines:
-    metrics:
-      receivers: [otlp]
-      processors: [batch]
-      exporters: [prometheus]
